+++ internal/provider/platform_resource.go
@@ -16,6 +16,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int64default"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/listdefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/setdefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringdefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
@@ -213,6 +214,7 @@ func (r *platformResource) Schema(_ context.Context, _ resource.SchemaRequest, r
 									stringvalidator.OneOf("PUBLISHED", "UNPUBLISHED"),
 								},
 							},
+							// TODO: check this is not empty if set to restricted and that it's set to the owner if private
 							"restricted_to_workspaces": schema.ListAttribute{
 								MarkdownDescription: "If the restriction is set to `RESTRICTED`, you can specify the workspace identifiers this meshPlatform is restricted to.",
 								ElementType:         types.StringType,
@@ -222,7 +224,7 @@ func (r *platformResource) Schema(_ context.Context, _ resource.SchemaRequest, r
 							},
 						},
 					},
-					"quota_definitions": schema.ListAttribute{
+					"quota_definitions": schema.SetAttribute{
 						MarkdownDescription: "List of quota definitions for the platform.",
 						Optional:            true,
 						Computed:            true,
@@ -230,7 +232,7 @@ func (r *platformResource) Schema(_ context.Context, _ resource.SchemaRequest, r
 						ElementType: types.ObjectType{
 							AttrTypes: quotaDefinitionAttrTypes,
 						},
-						Default: listdefault.StaticValue(types.ListValueMust(types.ObjectType{
+						Default: setdefault.StaticValue(types.SetValueMust(types.ObjectType{
 							AttrTypes: quotaDefinitionAttrTypes,
 						}, []attr.Value{})),
 					},
@@ -286,6 +288,21 @@ func meteringProcessingConfigSchema() schema.Attribute {
 	}
 }
 
+func secretEmbeddedSchema(description string, optional bool) schema.Attribute {
+	return schema.SingleNestedAttribute{
+		MarkdownDescription: description,
+		Required:            !optional,
+		Optional:            optional,
+		Attributes: map[string]schema.Attribute{
+			"plaintext": schema.StringAttribute{
+				MarkdownDescription: "Plaintext secret value",
+				Required:            true,
+				Sensitive:           true,
+			},
+		},
+	}
+}
+
 func (r *platformResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	platform := client.MeshPlatformCreate{
 		Metadata: client.MeshPlatformCreateMetadata{},
@@ -294,7 +311,6 @@ func (r *platformResource) Create(ctx context.Context, req resource.CreateReques
 	// Retrieve values from plan
 	resp.Diagnostics.Append(req.Plan.GetAttribute(ctx, path.Root("api_version"), &platform.ApiVersion)...)
 	resp.Diagnostics.Append(req.Plan.GetAttribute(ctx, path.Root("spec"), &platform.Spec)...)
-
 	resp.Diagnostics.Append(req.Plan.GetAttribute(ctx, path.Root("metadata").AtName("name"), &platform.Metadata.Name)...)
 	resp.Diagnostics.Append(req.Plan.GetAttribute(ctx, path.Root("metadata").AtName("owned_by_workspace"), &platform.Metadata.OwnedByWorkspace)...)
 
@@ -316,7 +332,7 @@ func (r *platformResource) Create(ctx context.Context, req resource.CreateReques
 		return
 	}
 
-	resp.Diagnostics.Append(resp.State.Set(ctx, createdPlatform)...)
+	resp.Diagnostics.Append(resp.State.Set(ctx, &createdPlatform)...)
 }
 
 func (r *platformResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {